Twiiter_Scrapping/requirements.txt
# filepath: f:\eFaida\Twiiter_Scrapping\requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
colorama==0.4.6
lxml==4.9.3  # Optional but makes BeautifulSoup faster
